[OSGi] Resolved all split package situations.
