+++ plc4j/drivers/ads/src/test/java/org/apache/plc4x/java/ads/connection/AdsSerialPlcConnectionTest.java
@@ -30,8 +30,8 @@ import org.apache.plc4x.java.ads.api.serial.types.*;
 import org.apache.plc4x.java.api.messages.PlcReadRequest;
 import org.apache.plc4x.java.api.messages.PlcReadResponse;
 import org.apache.plc4x.java.base.connection.NettyPlcConnection;
-import org.apache.plc4x.java.base.connection.SerialChannelFactory;
-import org.apache.plc4x.java.base.connection.SerialSocketAddress;
+import org.apache.plc4x.java.serial.connection.connection.SerialChannelFactory;
+import org.apache.plc4x.java.serial.connection.connection.SerialSocketAddress;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
