+++ plc4j/drivers/modbus/src/test/java/org/apache/plc4x/java/modbus/connection/ModbusSerialPlcConnectionTest.java
@@ -26,8 +26,8 @@ import org.apache.commons.lang3.reflect.MethodUtils;
 import org.apache.plc4x.java.api.messages.PlcReadRequest;
 import org.apache.plc4x.java.api.messages.PlcReadResponse;
 import org.apache.plc4x.java.base.connection.NettyPlcConnection;
-import org.apache.plc4x.java.base.connection.SerialChannelFactory;
-import org.apache.plc4x.java.base.connection.SerialSocketAddress;
+import org.apache.plc4x.java.serial.connection.connection.SerialChannelFactory;
+import org.apache.plc4x.java.serial.connection.connection.SerialSocketAddress;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;
