+++ plc4j/protocols/driver-bases/test/src/main/java/org/apache/plc4x/java/mock/connection/tcp/TcpHexDumper.java
@@ -16,7 +16,7 @@
  specific language governing permissions and limitations
  under the License.
  */
-package org.apache.plc4x.java.base.connection.tcp;
+package org.apache.plc4x.java.mock.connection.tcp;
 
 import org.apache.commons.io.HexDump;
 import org.apache.commons.io.IOUtils;
