+++ plc4j/protocols/driver-bases/test/src/main/java/org/apache/plc4x/java/mock/protocol/Plc4XSupportedDataTypes.java
@@ -16,7 +16,7 @@
  specific language governing permissions and limitations
  under the License.
  */
-package org.apache.plc4x.java.base.protocol;
+package org.apache.plc4x.java.mock.protocol;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.SerializationUtils;
