+++ plc4j/protocols/modbus/src/test/java/org/apache/plc4x/java/modbus/netty/Plc4XModbusProtocolTest.java
@@ -28,7 +28,7 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.plc4x.java.api.messages.PlcResponse;
 import org.apache.plc4x.java.base.messages.*;
-import org.apache.plc4x.java.base.protocol.Plc4XSupportedDataTypes;
+import org.apache.plc4x.java.mock.protocol.Plc4XSupportedDataTypes;
 import org.apache.plc4x.java.modbus.util.ModbusPlcFieldHandler;
 import org.junit.Before;
 import org.junit.Rule;
@@ -45,8 +45,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.apache.plc4x.java.base.protocol.Plc4XSupportedDataTypes.streamOfBigEndianDataTypePairs;
-import static org.apache.plc4x.java.base.util.Assert.assertByteEquals;
+import static org.apache.plc4x.java.mock.protocol.Plc4XSupportedDataTypes.streamOfBigEndianDataTypePairs;
+import static org.apache.plc4x.java.mock.util.Assert.assertByteEquals;
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assume.assumeThat;
